ケーブルのタッチ判定の修正

+++ app/src/main/java/com/example/testapplication/Device.java
@@ -8,9 +8,9 @@ public interface Device {
 
     void setHostname(TextView hostname);
 
-    int getCenterX();
+    float getCenterX();
 
-    int getCenterY();
+    float getCenterY();
 
     boolean getIsSet();
