+++ app/src/main/java/com/example/testapplication/Host.java
@@ -7,12 +7,12 @@ import android.widget.TextView;
 import androidx.appcompat.widget.AppCompatImageView;
 
 public class Host extends AppCompatImageView implements Device {
-    private int x, y; // 画像の中心座標
+    private float x, y; // 画像の中心座標
     private int host_ID; // ホストの番号
     private boolean isSet = false; // 設定済かどうか
     private TextView hostname;
 
-    public Host(Context context, int x, int y, int host_ID) {
+    public Host(Context context, float x, float y, int host_ID) {
         super(context);
         this.x = x;
         this.y = y;
@@ -38,11 +38,11 @@ public class Host extends AppCompatImageView implements Device {
         this.hostname = hostname;
     }
 
-    public int getCenterX() {
+    public float getCenterX() {
         return x;
     }
 
-    public int getCenterY() {
+    public float getCenterY() {
         return y;
     }
